requirements.txt
pandas
numpy
seaborn
matplotlib
tensorflow
keras
streamlit
scikit-learn
pillow
#-e .
